requirements_app.txt
# Python 12
# pip install -r requirements_app.txt

tensorflow~=2.17.0
mediapipe
opencv-python~=4.10.0.84
matplotlib~=3.9.2
scikit-learn~=1.5.2
pandas~=2.2.2
numpy~=1.26.4
pillow~=10.4.0
split-folders
gdown